Organize all the Python code to manage module and packages for the implementation
